move tests into root dir
