--- tests/go.mod
@@ -1,11 +1,0 @@
-module github.com/vechain/go-ecvrf/tests
-
-go 1.12
-
-require (
-	github.com/btcsuite/btcd v0.20.1-beta
-	github.com/stretchr/testify v1.5.1
-	github.com/vechain/go-ecvrf v0.0.0-20200305101714-4252ed3a3b96
-)
-
-replace github.com/vechain/go-ecvrf => ../
